Split bracket matching off into a separate module
Also split that code into functions
